tests,macvtap: Convert tests to use binding plugin
Following passt network binding plugin tests example
refactor macvtap tests as follows:

- Leave one test at macvtap.go to test the old macvtap
  interface API.
  Inline VM creation code to avoid coupling with
  bindingplugin_macvtap.go helpers.

- The rest of the tests are moved to bindingplugin_mavtap.go
  and changed to use the network binding plugins API to set
  macvtap interfaces.
